♻️ refactor: 공통 컴포넌트 경로 추가 변경
