+++ src/application/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from "../../common/hooks";
 export * from "./api";
 export * from "./domain";
